Move `ActivationSelector` to `activator\selection`
